tests/wpt/web-platform-tests/permissions-policy/private-state-token-issue-supported-by-permissions-policy.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(() => {
    assert_in_array('private-state-token-issuance', document.featurePolicy.features());
  }, 'document.featurePolicy.features should advertise private-state-token-issuance.');
</script>
